+++ paq8pxd.cpp
@@ -149,6 +149,8 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "models/im1bit.hpp"
 #include "models/sparsey.hpp"
 #include "models/distance.hpp"
+#include "models/im4bit.hpp"
+#include "models/im8bit.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -3050,47 +3052,6 @@ public:
 
 
 
-
- 
-inline U8 Clamp4(const int Px, const U8 n1, const U8 n2, const U8 n3, const U8 n4){
-  int maximum=n1;if(maximum<n2)maximum=n2;if(maximum<n3)maximum=n3;if(maximum<n4)maximum=n4;
-  int minimum=n1;if(minimum>n2)minimum=n2;if(minimum>n3)minimum=n3;if(minimum>n4)minimum=n4;
-  if(Px<minimum)return minimum;
-  if(Px>maximum)return maximum;
-  return Px;
-}
-inline U8 LogMeanDiffQt(const U8 a, const U8 b, const U8 limit = 7){
-  if (a==b) return 0;
-  U8 sign=a>b?8:0;
-  return sign | min(limit, ilog2((a+b)/max(2,abs(a-b)*2)+1));
-}
-inline U32 LogQt(const U8 Px, const U8 bits){
-  return (U32(0x100|Px))>>max(0,(int)(ilog2(Px)-bits));
-}
-class RingBuffer {
-  Array<U8> b;
-  U32 offset;
-public:
-  RingBuffer(const int i=0): b(i), offset(0) {}
-  void Fill(const U8 B) {
-    memset(&b[0], B, b.size());
-  }
-  void Add(const U8 B){
-    b[offset&(b.size()-1)] = B;
-    offset++;
-  }
-  int operator()(const int i) const {
-    return b[(offset-i)&(b.size()-1)];
-  }
-};
-
-inline U8 Paeth(U8 W, U8 N, U8 NW){
-  int p = W+N-NW;
-  int pW=abs(p-(int)W), pN=abs(p-(int)N), pNW=abs(p-(int)NW);
-  if (pW<=pN && pW<=pNW) return W;
-  else if (pN<=pNW) return N;
-  return NW;
-}
 //////////////////////////// im24bitModel /////////////////////////////////
 // Model for 24-bit image data
 
@@ -3647,553 +3608,7 @@ inline int sqrbuf(int i) {
  
 };
 
-//////////////////////////// im8bitModel /////////////////////////////////
-// Model for 8-bit image data
 
-class im8bitModel1: public Model {
-  typedef enum {
-    nOLS = 5,
-    nMaps0 = 2,
-    nMaps1 = 55,
-    nMaps = 62,  //nMaps0 + nMaps1 + nOLS
-    nPltMaps = 4
- } im8M;
- int inpts;
- ContextMap2 cm;
- int col;
- BlockData& xx;
- Buf& buf;
- int ctx, lastPos, lastWasPNG, line, x, filter, gray,isPNG,jump;
- int framePos, prevFramePos, frameWidth, prevFrameWidth;
- U32& c4;
- int& c0;
- int& bpos;
- StationaryMap Map[nMaps] = {     0, {15,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1},
-                                     {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1},
-                                     {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1},
-                                     {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1},
-                                     {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1},
-                                     {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1},
-                                     {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1},
-                                     {11,1}, {11,1}, {11,1}, {11,1}, {11,1}, {11,1} };
- SmallStationaryContextMap pltMap[nPltMaps] = { {11,1},{11,1},{11,1},{11,1} };
- IndirectMap sceneMap[5]{ {8}, {8}, {22,1}, {11,1}, {11,1} };
- IndirectContext<U8> iCtx[nPltMaps] = { {16,8}, {16,8}, {16,8}, {16,8} };
- U8 px , res,prvFrmPx , prvFrmPred ;// current PNG filter prediction, expected residual
- RingBuffer buffer;// internal rotating buffer for PNG unfiltered pixel data
- bool filterOn;
- int columns[2] ={ 1,1 }  , column[2]={ 1,1 } ;
- Array<short> jumps;
- U8 WWWWWW, WWWWW, WWWW, WWW, WW, W;
-    U8 NWWWW, NWWW, NWW, NW, N, NE, NEE, NEEE, NEEEE;
-    U8 NNWWW, NNWW, NNW, NN, NNE, NNEE, NNEEE;
-    U8 NNNWW, NNNW, NNN, NNNE, NNNEE;
-    U8 NNNNW, NNNN, NNNNE;
-    U8 NNNNN;
-    U8 NNNNNN;
-    Array<U8>    MapCtxs ;
-    Array<U8>    pOLS;
-    const double lambda[nOLS] ={ 0.996, 0.87, 0.93, 0.8, 0.9 };
-    const int num[nOLS] ={ 32, 12, 15, 10, 14 };
-    OLS<double, U8> ols[nOLS] = { 
-    {num[0], 1, lambda[0]},
-    {num[1], 1, lambda[1]},
-    {num[2], 1, lambda[2]},
-    {num[3], 1, lambda[3]},
-    {num[4], 1, lambda[4]}
-  };
-  OLS<double, U8> sceneOls;
-    const U8 *ols_ctx1[32] = { &WWWWWW, &WWWWW, &WWWW, &WWW, &WW, &W, &NWWWW, &NWWW, &NWW, &NW, &N, &NE, &NEE, &NEEE, &NEEEE, &NNWWW, &NNWW, &NNW, &NN, &NNE, &NNEE, &NNEEE, &NNNWW, &NNNW, &NNN, &NNNE, &NNNEE, &NNNNW, &NNNN, &NNNNE, &NNNNN, &NNNNNN };
-    const U8 *ols_ctx2[12] = { &WWW, &WW, &W, &NWW, &NW, &N, &NE, &NEE, &NNW, &NN, &NNE, &NNN }; 
-    const U8 *ols_ctx3[15] = { &N, &NE, &NEE, &NEEE, &NEEEE, &NN, &NNE, &NNEE, &NNEEE, &NNN, &NNNE, &NNNEE, &NNNN, &NNNNE, &NNNNN };
-    const U8 *ols_ctx4[10] = { &N, &NE, &NEE, &NEEE, &NN, &NNE, &NNEE, &NNN, &NNNE, &NNNN };
-    const U8 *ols_ctx5[14] = { &WWWW, &WWW, &WW, &W, &NWWW, &NWW, &NW, &N, &NNWW, &NNW, &NN, &NNNW, &NNN, &NNNN };
-    const U8 **ols_ctxs[nOLS] = { &ols_ctx1[0], &ols_ctx2[0], &ols_ctx3[0], &ols_ctx4[0], &ols_ctx5[0] };
-  
-public:
-  im8bitModel1( BlockData& bd):  inpts(49+nPltMaps),cm(CMlimit(MEM()*4), inpts,M_IM8,
-  CM_RUN1+
-  CM_RUN0+
-  CM_MAIN1+
-  CM_MAIN2+
-  CM_MAIN3+
-  CM_MAIN4+
-  CM_M12
-  ),col(0),xx(bd),buf(bd.buf), ctx(0),lastPos(0), lastWasPNG(0),line(0), x(0),
-   filter(0),gray(0),isPNG(0),jump(0), framePos(0), prevFramePos(0), frameWidth(0), prevFrameWidth(0), c4(bd.c4),c0(bd.c0),bpos(bd.bpos), px(0),prvFrmPx(0), prvFrmPred(0),
-   res (0),buffer(0x100000),filterOn(false),jumps(0x8000),WWWWWW(0), WWWWW(0), WWWW(0), WWW(0), WW(0), W(0),
-      NWWWW(0), NWWW(0), NWW(0), NW(0), N(0), NE(0), NEE(0), NEEE(0), NEEEE(0),
-      NNWWW(0), NNWW(0), NNW(0), NN(0), NNE(0), NNEE(0), NNEEE(0),
-      NNNWW(0), NNNW(0), NNN(0), NNNE(0), NNNEE(0),
-      NNNNW(0), NNNN(0), NNNNE(0), NNNNN(0), NNNNNN(0),MapCtxs(nMaps1),   pOLS(nOLS),sceneOls(13, 1, 0.994){
-
-  }
-  int inputs() {return inpts*cm.inputs()+nMaps*2+nPltMaps*2+5*2;}
-  int nets() {return ( 2048+5)+    6*16+    6*32+    256+    1024+    64+    128+    256;}
-  int netcount() {return 8;}
-  
-int p(Mixer& m,int w,int val2=0){
-  assert(w>3); 
-  if (!bpos) {
-    if (xx.blpos==1){
-      isPNG=  (xx.filetype==PNG8?1:xx.filetype==PNG8GRAY?1:0);
-      gray=xx.filetype==PNG8GRAY?1:xx.filetype==IMAGE8GRAY?1:0;
-      x =0; line = jump =  px= 0;
-      filterOn = false;
-      columns[0] = max(1,w/max(1,ilog2(w)*2));
-      columns[1] = max(1,columns[0]/max(1,ilog2(columns[0])));
-      if (gray){
-        if (lastPos && lastWasPNG!=isPNG){
-          for (int i=0;i<nMaps;i++)
-            Map[i].Reset();
-            xx.count=0;
-        }
-        lastWasPNG = isPNG;
-      }
-      buffer.Fill(0x7F);
-      prevFramePos = framePos;
-      framePos = xx.blpos;
-      prevFrameWidth = frameWidth;
-      frameWidth = w;
-    }
-    else{
-      x++;
-      if(x>=w+isPNG){x=0;line++;}
-    }
-//lastPos = xx.blpos;
-    if (isPNG){
-      if (x==1)
-        filter = (U8)c4;
-      else{
-        U8 B = (U8)c4;
-
-        switch (filter){
-          case 1: {
-            buffer.Add((U8)( B + buffer(1)*(x>2 || !x) ) );
-            filterOn = x>1;
-            px = buffer(1);
-            break;
-          }
-          case 2: {
-            buffer.Add((U8)( B + buffer(w)*(filterOn=(line>0)) ) );
-            px = buffer(w);
-            break;
-          }
-          case 3: {
-            buffer.Add((U8)( B + (buffer(w)*(line>0) + buffer(1)*(x>2 || !x))/2 ) );
-            filterOn = (x>1 || line>0);
-            px = (buffer(1)*(x>1)+buffer(w)*(line>0))/2;
-            break;
-          }
-          case 4: {
-            buffer.Add((U8)( B + Paeth(buffer(1)*(x>2 || !x), buffer(w)*(line>0), buffer(w+1)*(line>0 && (x>2 || !x))) ) );
-            filterOn = (x>1 || line>0);
-            px = Paeth(buffer(1)*(x>1),buffer(w)*(line>0),buffer(w+1)*(x>1 && line>0));
-            break;
-          }
-          default: buffer.Add(B);
-            filterOn = false;
-            px = 0;
-        }
-        if(!filterOn)px=0;
-      }
-    }  
-    else {
-      buffer.Add((U8)c4);
-      if (x==0) {
-        memset(&jumps[0], 0, sizeof(short)*jumps.size());
-        if (line>0 && w>8) {
-          U8 bMask = 0xFF-((1<<gray)-1);
-          U32 pMask = bMask*0x01010101u;
-          U32 left=0, right=0;
-          int l=min(w, (int)jumps.size()), end=l-4;
-          do {
-            left = ((buffer(l-x)<<24)|(buffer(l-x-1)<<16)|(buffer(l-x-2)<<8)|buffer(l-x-3))&pMask;
-            int i = end;
-            while (i>=x+4) {
-              right = ((buffer(l-i-3)<<24)|(buffer(l-i-2)<<16)|(buffer(l-i-1)<<8)|buffer(l-i))&pMask;
-              if (left==right) {
-                int j=(i+3-x-1)/2, k=0;
-                for (; k<=j; k++) {
-                  if (k<4 || (buffer(l-x-k)&bMask)==(buffer(l-i-3+k)&bMask)) {
-                    jumps[x+k] = -(x+(l-i-3)+2*k);
-                    jumps[i+3-k] = i+3-x-2*k;
-                  }
-                  else
-                    break;
-                }
-                x+=k;
-                end-=k;
-                break;
-              }
-              i--;
-            }
-            x++;
-            if (x>end)
-              break;
-          } while (x+4<l);
-          x = 0;
-        }
-      }
-    }
-
-    if (x || !isPNG){
-      column[0]=(x-isPNG)/columns[0];
-      column[1]=(x-isPNG)/columns[1];
-      
-      WWWWW=buffer(5), WWWW=buffer(4), WWW=buffer(3), WW=buffer(2), W=buffer(1);
-      NWWWW=buffer(w+4), NWWW=buffer(w+3), NWW=buffer(w+2), NW=buffer(w+1), N=buffer(w), NE=buffer(w-1), NEE=buffer(w-2), NEEE=buffer(w-3), NEEEE=buffer(w-4);
-      NNWWW=buffer(w*2+3), NNWW=buffer(w*2+2), NNW=buffer(w*2+1), NN=buffer(w*2), NNE=buffer(w*2-1), NNEE=buffer(w*2-2), NNEEE=buffer(w*2-3);
-      NNNWW=buffer(w*3+2), NNNW=buffer(w*3+1), NNN=buffer(w*3), NNNE=buffer(w*3-1), NNNEE=buffer(w*3-2);
-      NNNNW=buffer(w*4+1), NNNN=buffer(w*4), NNNNE=buffer(w*4-1);
-      NNNNN=buffer(w*5);
-      NNNNNN=buffer(w*6);
-      if (prevFramePos>0 && prevFrameWidth==w){
-        int offset = prevFramePos+line*w+x;
-        prvFrmPx = buf[offset];
-        if (gray) {
-          sceneOls.Update(W);
-          sceneOls.Add(W); sceneOls.Add(NW); sceneOls.Add(N); sceneOls.Add(NE);
-          for (int i=-1; i<2; i++) {
-            for (int j=-1; j<2; j++)
-              sceneOls.Add(buf[offset+i*w+j]);
-          }
-          prvFrmPred = Clip(int(floor(sceneOls.Predict())));
-        }
-        else
-          prvFrmPred = W;
-      }
-      else
-        prvFrmPx = prvFrmPred = W;
-      sceneMap[0].set_direct(prvFrmPx);
-      sceneMap[1].set_direct(prvFrmPred);
-
-      int j = 0;
-      jump = jumps[min(x,(int)jumps.size()-1)];
-
-      U64 i= (filterOn ? (filter+1)*64 : 0) + (gray*1024);
-      cm.set(hash(++i, (jump!=0)?(0x100|buffer(abs(jump)))*(1-2*(jump<0)):N, line&3));
-      
-      
-      if (!gray){
-       for (j=0; j<nPltMaps; j++)
-        iCtx[j]+=W;
-      iCtx[0]=W|(NE<<8);
-      iCtx[1]=W|(N<<8);
-      iCtx[2]=W|(WW<<8);
-      iCtx[3]=N|(NN<<8);
-      
-        cm.set(hash(++i, W, px));
-        cm.set(hash(++i, W, px, column[0]));
-        cm.set(hash(++i, N, px));
-        cm.set(hash(++i, N, px, column[0]));
-        cm.set(hash(++i, NW, px));
-        cm.set(hash(++i, NW, px, column[0]));
-        cm.set(hash(++i, NE, px));
-        cm.set(hash(++i, NE, px, column[0]));
-        cm.set(hash(++i, NWW, px));
-        cm.set(hash(++i, NEE, px));
-        cm.set(hash(++i, WW, px));
-        cm.set(hash(++i, NN, px));
-        cm.set(hash(++i, W, N, px));
-        cm.set(hash(++i, W, NW, px));
-        cm.set(hash(++i, W, NE, px));
-        cm.set(hash(++i, W, NEE, px));
-        cm.set(hash(++i, W, NWW, px));
-        cm.set(hash(++i, N, NW, px));
-        cm.set(hash(++i, N, NE, px));
-        cm.set(hash(++i, NW, NE, px));
-        cm.set(hash(++i, W, WW, px));
-        cm.set(hash(++i, N, NN, px));
-        cm.set(hash(++i, NW, NNWW, px));
-        cm.set(hash(++i, NE, NNEE, px));
-        cm.set(hash(++i, NW, NWW, px));
-        cm.set(hash(++i, NW, NNW, px));
-        cm.set(hash(++i, NE, NEE, px));
-        cm.set(hash(++i, NE, NNE, px));
-        cm.set(hash(++i, N, NNW, px));
-        cm.set(hash(++i, N, NNE, px));
-        cm.set(hash(++i, N, NNN, px));
-        cm.set(hash(++i, W, WWW, px));
-        cm.set(hash(++i, WW, NEE, px));
-        cm.set(hash(++i, WW, NN, px));
-        cm.set(hash(++i, W, buffer(w-3), px));
-        cm.set(hash(++i, W, buffer(w-4), px));
-        cm.set(hash(++i, W, N,NW, px));
-        cm.set(hash(++i, N, NN,NNN, px));
-        cm.set(hash(++i, W, NE,NEE, px));
-        cm.set(hash(hash( W,NW,N,NE), px));
-        cm.set(hash( hash(N,NE,NN,NNE), px));
-        cm.set(hash( hash(N,NW,NNW,NN), px));
-        cm.set(hash( hash(W,WW,NWW,NW), px));
-        cm.set(hash(++i, W, NW<<8 | N, WW<<8 | NWW, px));
-        cm.set(hash(++i, px, column[0]));
-        cm.set(hash(++i, px));
-        cm.set(hash(++i, N, px, column[1] ));
-        cm.set(hash(++i, W, px, column[1] ));
-        for (int j=0; j<nPltMaps; j++)
-          cm.set(hash(++i, iCtx[j](), px));
-        ctx = min(0x1F,(x-isPNG)/min(0x20,columns[0]));
-        res = W;
-      }
-      else{
-      MapCtxs[j++] = Clamp4(W+N-NW,W,NW,N,NE);
-      MapCtxs[j++] = Clip(W+N-NW);
-      MapCtxs[j++] = Clamp4(W+NE-N,W,NW,N,NE);
-      MapCtxs[j++] = Clip(W+NE-N);
-      MapCtxs[j++] = Clamp4(N+NW-NNW,W,NW,N,NE);
-      MapCtxs[j++] = Clip(N+NW-NNW);
-      MapCtxs[j++] = Clamp4(N+NE-NNE,W,N,NE,NEE);
-      MapCtxs[j++] = Clip(N+NE-NNE);
-      MapCtxs[j++] = (W+NEE)/2;
-      MapCtxs[j++] = Clip(N*3-NN*3+NNN);
-      MapCtxs[j++] = Clip(W*3-WW*3+WWW);
-      MapCtxs[j++] = (W+Clip(NE*3-NNE*3+buffer(w*3-1)))/2;
-      MapCtxs[j++] = (W+Clip(NEE*3-buffer(w*2-3)*3+buffer(w*3-4)))/2;
-      MapCtxs[j++] = Clip(NN+buffer(w*4)-buffer(w*6));
-      MapCtxs[j++] = Clip(WW+buffer(4)-buffer(6));
-      MapCtxs[j++] = Clip((buffer(w*5)-6*buffer(w*4)+15*NNN-20*NN+15*N+Clamp4(W*2-NWW,W,NW,N,NN))/6);
-      MapCtxs[j++] = Clip((-3*WW+8*W+Clamp4(NEE*3-NNEE*3+buffer(w*3-2),NE,NEE,buffer(w-3),buffer(w-4)))/6);
-      MapCtxs[j++] = Clip(NN+NW-buffer(w*3+1));
-      MapCtxs[j++] = Clip(NN+NE-buffer(w*3-1));
-      MapCtxs[j++] = Clip((W*2+NW)-(WW+2*NWW)+buffer(w+3));
-      MapCtxs[j++] = Clip(((NW+NWW)/2)*3-buffer(w*2+3)*3+(buffer(w*3+4)+buffer(w*3+5))/2);
-      MapCtxs[j++] = Clip(NEE+NE-buffer(w*2-3));
-      MapCtxs[j++] = Clip(NWW+WW-buffer(w+4));
-      MapCtxs[j++] = Clip(((W+NW)*3-NWW*6+buffer(w+3)+buffer(w*2+3))/2);
-      MapCtxs[j++] = Clip((NE*2+NNE)-(NNEE+buffer(w*3-2)*2)+buffer(w*4-3));
-      MapCtxs[j++] = buffer(w*6);
-      MapCtxs[j++] = (buffer(w-4)+buffer(w-6))/2;
-      MapCtxs[j++] = (buffer(4)+buffer(6))/2;
-      MapCtxs[j++] = (W+N+buffer(w-5)+buffer(w-7))/4;
-      MapCtxs[j++] = Clip(buffer(w-3)+W-NEE);
-      MapCtxs[j++] = Clip(4*NNN-3*buffer(w*4));
-      MapCtxs[j++] = Clip(N+NN-NNN);
-      MapCtxs[j++] = Clip(W+WW-WWW);
-      MapCtxs[j++] = Clip(W+NEE-NE);
-      MapCtxs[j++] = Clip(WW+NEE-N);
-      MapCtxs[j++] = (Clip(W*2-NW)+Clip(W*2-NWW)+N+NE)/4;
-      MapCtxs[j++] = Clamp4(N*2-NN,W,N,NE,NEE);
-      MapCtxs[j++] = (N+NNN)/2;
-      MapCtxs[j++] = Clip(NN+W-NNW);
-      MapCtxs[j++] = Clip(NWW+N-NNWW);
-      MapCtxs[j++] = Clip((4*WWW-15*WW+20*W+Clip(NEE*2-NNEE))/10);
-      MapCtxs[j++] = Clip((buffer(w*3-3)-4*NNEE+6*NE+Clip(W*3-NW*3+NNW))/4);
-      MapCtxs[j++] = Clip((N*2+NE)-(NN+2*NNE)+buffer(w*3-1));
-      MapCtxs[j++] = Clip((NW*2+NNW)-(NNWW+buffer(w*3+2)*2)+buffer(w*4+3));
-      MapCtxs[j++] = Clip(NNWW+W-buffer(w*2+3));
-      MapCtxs[j++] = Clip((-buffer(w*4)+5*NNN-10*NN+10*N+Clip(W*4-NWW*6+buffer(w*2+3)*4-buffer(w*3+4)))/5);
-      MapCtxs[j++] = Clip(NEE+Clip(buffer(w-3)*2-buffer(w*2-4))-buffer(w-4));
-      MapCtxs[j++] = Clip(NW+W-NWW);
-      MapCtxs[j++] = Clip((N*2+NW)-(NN+2*NNW)+buffer(w*3+1));
-      MapCtxs[j++] = Clip(NN+Clip(NEE*2-buffer(w*2-3))-NNE);
-      MapCtxs[j++] = Clip((-buffer(4)+5*WWW-10*WW+10*W+Clip(NE*2-NNE))/5);
-      MapCtxs[j++] = Clip((-buffer(5)+4*buffer(4)-5*WWW+5*W+Clip(NE*2-NNE))/4);
-      MapCtxs[j++] = Clip((WWW-4*WW+6*W+Clip(NE*3-NNE*3+buffer(w*3-1)))/4);
-      MapCtxs[j++] = Clip((-NNEE+3*NE+Clip(W*4-NW*6+NNW*4-buffer(w*3+1)))/3);
-      MapCtxs[j++] = ((W+N)*3-NW*2)/4;
-      for (j=0; j<nOLS; j++) {
-        ols[j].Update(W);
-        pOLS[j] = Clip(int(floor(ols[j].Predict(ols_ctxs[j]))));
-      }
-      
-     
-        cm.set();
-        cm.set(hash(++i, N, px));
-        cm.set(hash(++i, N-px));
-        cm.set(hash(++i, W, px));
-        cm.set(hash(++i, NW, px));
-        cm.set(hash(++i, NE, px));
-        cm.set(hash(++i, N, NN, px));
-        cm.set(hash(++i, W, WW, px));
-        cm.set(hash(++i, NE, NNEE, px ));
-        cm.set(hash(++i, NW, NNWW, px ));
-        cm.set(hash(++i, W, NEE, px));
-        cm.set(hash(++i, (Clamp4(W+N-NW,W,NW,N,NE)-px)/2, LogMeanDiffQt(Clip(N+NE-NNE), Clip(N+NW-NNW))));
-        cm.set(hash(++i, (W-px)/4, (NE-px)/4, column[0]));
-        cm.set(hash(++i, (Clip(W*2-WW)-px)/4, (Clip(N*2-NN)-px)/4));
-        cm.set(hash(++i, (Clamp4(N+NE-NNE,W,N,NE,NEE)-px)/4, column[0]));
-        cm.set(hash(++i, (Clamp4(N+NW-NNW,W,NW,N,NE)-px)/4, column[0]));
-        cm.set(hash(++i, (W+NEE)/4, px, column[0]));
-        cm.set(hash(++i, Clip(W+N-NW)-px, column[0]));
-        cm.set(hash(++i, Clamp4(N*3-NN*3+NNN,W,N,NN,NE), px, LogMeanDiffQt(W,Clip(NW*2-NNW))));
-        cm.set(hash(++i, Clamp4(W*3-WW*3+WWW,W,N,NE,NEE), px, LogMeanDiffQt(N,Clip(NW*2-NWW))));
-        cm.set(hash(++i, (W+Clamp4(NE*3-NNE*3+NNNE,W,N,NE,NEE))/2, px, LogMeanDiffQt(N,(NW+NE)/2)));
-        cm.set(hash(++i, (N+NNN)/8, Clip(N*3-NN*3+NNN)/4, px));
-        cm.set(hash(++i, (W+WWW)/8, Clip(W*3-WW*3+WWW)/4, px));
-        cm.set(hash(++i, Clip((-buffer(4)+5*WWW-10*WW+10*W+Clamp4(NE*4-NNE*6+buffer(w*3-1)*4-buffer(w*4-1),N,NE,buffer(w-2),buffer(w-3)))/5)-px));
-        cm.set(hash(++i, Clip(N*2-NN)-px, LogMeanDiffQt(N,Clip(NN*2-NNN))));
-        cm.set(hash(++i, Clip(W*2-WW)-px, LogMeanDiffQt(NE,Clip(N*2-NW))));
-
-      
-        if (isPNG)
-          ctx = ((abs(W-N)>8)<<10)|((W>N)<<9)|((abs(N-NW)>8)<<8)|((N>NW)<<7)|((abs(N-NE)>8)<<6)|((N>NE)<<5)|((W>WW)<<4)|((N>NN)<<3)|min(5,filterOn?filter+1:0);
-        else
-          ctx = min(0x1F,x/max(1,w/min(32,columns[0])))|( ( ((abs(W-N)*16>W+N)<<1)|(abs(N-NW)>8) )<<5 )|((W+N)&0x180);
-
-        res = Clamp4(W+N-NW,W,NW,N,NE)-px;
-      }
-        xx.Image.pixels.W = W;
-        xx.Image.pixels.N = N;
-        xx.Image.pixels.NN = NN;
-        xx.Image.pixels.WW = WW;
-        xx.Image.ctx = ctx>>gray;
-    }
-  }
-  U8 B=(c0<<(8-bpos));
-  if (x || !isPNG){
-      if (gray) {
-    int i=1;
-    Map[i++].set((((U8)(Clip(W+N-NW)-px-B))*8+bpos)|(LogMeanDiffQt(Clip(N+NE-NNE),Clip(N+NW-NNW))<<11));
-    
-    for (int j=0; j<nMaps1; i++, j++)
-      Map[i].set((MapCtxs[j]-px-B)*8+bpos);
-
-    for (int j=0; i<nMaps; i++, j++)
-      Map[i].set((pOLS[j]-px-B)*8+bpos);
-  }
-    sceneMap[2].set_direct(finalize64(hash(x, line), 19)*8+bpos);
-    sceneMap[3].set_direct((prvFrmPx-B)*8+bpos);
-    sceneMap[4].set_direct((prvFrmPred-B)*8+bpos);
-}
-  // Predict next bit
-  if (x || !isPNG){
-  col=(col+1)&7;
-      if (val2==1)  return 1;   
-    cm.mix(m);
-    if (gray){
-      for (int i=0;i<nMaps;i++)
-        Map[i].mix1(m);
-    }else {
-      for (int i=0; i<nPltMaps; i++) {
-        pltMap[i].set((bpos<<8)|iCtx[i]());
-        pltMap[i].mix(m);
-      }
-    }
-    for (int i=0; i<5; i++)
-      sceneMap[i].mix(m, (prevFramePos>0 && prevFrameWidth==w), 4, 255);
-
-    m.set(5+ctx, 2048+5);
-    m.set(col*2+(isPNG && c0==((0x100|res)>>(8-bpos))) + min(5,filterOn?filter+1:0)*16, 6*16);
-    m.set(((isPNG?px:N+W)>>4) + min(5,filterOn?filter+1:0)*32, 6*32);
-    m.set(c0, 256);
-    m.set( ((abs((int)(W-N))>4)<<9)|((abs((int)(N-NE))>4)<<8)|((abs((int)(W-NW))>4)<<7)|((W>N)<<6)|((N>NE)<<5)|((W>NW)<<4)|((W>WW)<<3)|((N>NN)<<2)|((NW>NNWW)<<1)|(NE>NNEE), 1024 );
-    m.set(min(63,column[0]), 64);
-    m.set(min(127,column[1]), 128);
-    m.set( min(255,(x+line)/32), 256);
-  }
-  else{
-    m.add( -2048+((filter>>(7-bpos))&1)*4096 );
-    m.set(min(4,filter),5);
-  }
-  return 0; //8 8 32 256 512 1792
-  }
-  // Square buf(i)
-inline int sqrbuf(int i) {
-  assert(i>0);
-  return buf(i)*buf(i);
-}
-  virtual ~im8bitModel1(){ }
- 
-};
-
-//////////////////////////// im4bitModel /////////////////////////////////
-
-// Model for 4-bit image data
-class im4bitModel1: public Model {
-    BlockData& x;
-    Buf& buf;
-    BH<16> t;
-    const int S; // number of contexts
-    Array<U8*> cp;
-    StateMap *sm;
-    StateMap map;
-    U8 WW, W, NWW, NW, N, NE, NEE, NNWW, NNW, NN, NNE, NNEE;
-    int col, line, run, prevColor, px;
-    public:
- im4bitModel1( BlockData& bd,U32 val=0 ): x(bd),buf(bd.buf),t( CMlimit((level>14?MEM()/2:MEM())/8) ),S(14),cp(S),map(16), WW(0), W(0), NWW(0), NW(0), N(0), NE(0),
-  NEE(0), NNWW(0), NNW(0), NN(0), NNE(0), NNEE(0),col(0), line(0), run(0), prevColor(0), px(0) {
-   sm=new StateMap[S];
-   for (int i=0;i<S;i++)
-      cp[i]=t[263*i]+1;
-   }
-  int inputs() {return S*3+2;}
-  int nets() {return 256+   512+   512+  1024+   16+  1;}
-  int netcount() {return 6;}
-   
-int p(Mixer& m,int w=0,int val2=0)  {
-  int i;
-  if (x.blpos==1){//helps only on bigger files+1024kb
-      //t.reset();
-      for (i=0;i<S;i++)
-      cp[i]=t[263*i]+1;
-  }
-  for (i=0;i<S;i++)
-    *cp[i]=nex(*cp[i],x.y);
-
-  if (x.bpos==0 || x.bpos==4){
-      WW=W, NWW=NW, NW=N, N=NE, NE=NEE, NNWW=NWW, NNW=NN, NN=NNE, NNE=NNEE;
-      if (x.bpos==0)
-        {
-        W=x.c4&0xF, NEE=buf(w-1)>>4, NNEE=buf(w*2-1)>>4;}
-      else
-       {
-        W=x.c0&0xF, NEE=buf(w-1)&0xF, NNEE=buf(w*2-1)&0xF; }
-      run=(W!=WW || col==0)?(prevColor=WW,0):min(0xFFF,run+1);
-      px=1, i=0;
-
-      cp[i++]=t[hash(W,NW,N)]+1;
-      cp[i++]=t[hash(N, min(0xFFF, col/8))]+1;
-      cp[i++]=t[hash(W,NW,N,NN,NE)]+1;
-      cp[i++]=t[hash(W, N, NE+NNE*16, NEE+NNEE*16)]+1;
-      cp[i++]=t[hash(W, N, NW+NNW*16, NWW+NNWW*16)]+1;
-      cp[i++]=t[hash(W, ilog2(run+1), prevColor, col/max(1,w/2) )]+1;
-      cp[i++]=t[hash(NE, min(0x3FF, (col+line)/max(1,w*8)))]+1;
-      cp[i++]=t[hash(NW, (col-line)/max(1,w*8))]+1;
-      cp[i++]=t[hash(WW*16+W,NN*16+N,NNWW*16+NW)]+1;
-      cp[i++]=t[hash(i,N,NN)]+1;
-      cp[i++]=t[hash(i,W,WW)]+1;
-      cp[i++]=t[hash(i,W,NE)]+1;
-      cp[i++]=t[hash(i,WW,NN,NEE)]+1;
-      cp[i++]=t[-1]+1;
-      
-      col++;
-      if(col==w*2){col=0;line++;}
-        x.Image.pixels.W = W;
-        x.Image.pixels.N = N;
-        x.Image.pixels.NN = NN;
-        x.Image.pixels.WW = WW;
-        x.Image.ctx = W*16+px;
-  }
-  else{
-    px+=px+x.y;
-    int j=(x.y+1)<<(x.bpos&3);
-    for (i=0;i<S;i++)
-      cp[i]+=j;
-  }
-  // predict
-  for (int i=0; i<S; i++) {
-    const U8 s = *cp[i];
-    const int n0=-!nex(s, 2), n1=-!nex(s, 3);
-    const int p1 = sm[i].p1(s,x.y);
-    const int st = stretch(p1)>>1;
-    m.add(st);
-    m.add((p1-2048)>>3);
-    m.add(st*abs(n1-n0));
-  }
-  m.add(stretch(map.p(px,x.y)));
- 
-  m.set(W*16+px, 256);
-  m.set(min(31,col/max(1,w/16))+N*32, 512);
-  m.set((x.bpos&3)+4*W+64*min(7,ilog2(run+1)), 512);
-  m.set(W+NE*16+(x.bpos&3)*256, 1024);
-  m.set(px, 16);
-  m.set(0,1);
-  return 0;
-}
- virtual ~im4bitModel1(){  delete[] sm;}
- 
-};
